viton/parsing/requirements.txt
click==7.1.2
dataclasses==0.8
fastapi==0.68.0
h11==0.9.0
httptools==0.1.2
importlib-metadata==4.8.3
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter-tensorboard==0.1.10
kiwisolver==1.1.0
libarchive-c==2.8
lief==0.9.0
llvmlite==0.28.0
lmdb==0.97
MarkupSafe==1.1.1
maskrcnn-benchmark==0.1
matplotlib==3.1.1
mccabe==0.6.1
mlperf-compliance==0.0.10
more-itertools==7.2.0
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.0
numba==0.43.1
numpy==1.16.4
onnx==1.5.0
opencv-python==3.4.1.15
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pathvalidate==0.29.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow-SIMD==5.3.0.post1
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.12.0
preshed==2.0.1
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pycocotools==2.0.2
pycosat==0.6.3
pydantic==1.9.2
pydot==1.4.1
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.0
pytablewriter==0.46.1
pytest-pythonpath==0.7.3
python-dateutil==2.8.0
python-multipart==0.0.5
python-nvd3==0.15.0
python-slugify==3.0.3
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
regex==2018.1.10
revtok==0.0.3
ruamel-yaml==0.15.46
sacrebleu==1.2.10
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.9.0
soupsieve==1.9.2
SSD==0.1
starlette==0.14.2
tabledata==0.9.1
tabulate==0.8.3
tensorboard==1.14.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
toolz==0.10.0
torchtext==0.4.0
tornado==6.0.3
tqdm==4.31.1
traitlets==4.3.2
typepy==0.6.0
typing==3.7.4.1
typing-extensions==4.1.1
ujson==1.35
uvicorn==0.11.8
uvloop==0.14.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
Werkzeug==0.15.6
wrapt==1.10.11
yacs==0.1.6
zipp==0.6.0
